% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGitBookVersions.R
\name{getGitBookVersions}
\alias{getGitBookVersions}
\title{Get GitBook versions installed}
\usage{
getGitBookVersions()
}
\description{
Prints the GitBook versions installed in the system
}
\details{
This function is a system call to: \code{gitbook ls}.
To get the GitBook versions installed in the system.
}
