% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGitBookCli.R
\name{installGitBookCli}
\alias{installGitBookCli}
\title{Install npm gitbook-cli package}
\usage{
installGitBookCli()
}
\description{
Install the npm gitbook-cli package globally in the system
}
\details{
This function is a system call to: \code{npm install -g gitbook-cli}.
To install the npm gitbook package in the system.
}
