% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGitBookVersion.R
\name{installGitBookVersion}
\alias{installGitBookVersion}
\title{Install GitBook version}
\usage{
installGitBookVersion(version = "2.6.4")
}
\arguments{
\item{version}{GitBook version that the user wants to install (Default: 2.6.4).}
}
\description{
Install a GitBook version provided by the user. (Default: 2.6.4)
}
\details{
This function is a system call to: \code{gitbook fetch [version]}.
To install a GitBook version in the system.

Default version 2.6.4 because it lets a browser navigate between pages automatically.
}
